requirements.txt
fastapi
uvicorn
SQLAlchemy
pydantic
databases[postgresql]
psycopg2-binary
python-multipart
python-dotenv